Merge: Add MOTD option in updater-script.
This will create a message that is displayed while the user is flashing
the rom. To define the MOTD, just define 'PRODUCT_MOTD'. Currently this
is being used in the evervolv vendor for defining a seperate message
for each phone/device. As of right now the script splits the lines @
48 characters. You can also use \n for a new line.