build: changes for LINARO_BUILD
define LINARO_BUILD=1 to use linaro optimizations
optional dependencies of LINARO_BUILD:
LINARO_CPP11=1 to build with ISO CPP++11
    currently breaks the build thanks qcom!!
LINARO_OPT3=1 to build with -O3
    O3 breaks something and results in no display for passion, it also
    creates a build that is 9mb larger

Until the linaro toolchain is added to the tree we must specify
TARGET_TOOLS_PREFIX on the command line pointing to an external dl of
linaros gcc 4.7. So if that is not specified error out.

Change-Id: I0727c35b1947942f6b86322307c8390bbb798925