Fix case where boot/recovery.img were being built with wrong  params.
The boot and recovery images  now get built using the same params during ota package
generation as during a normal build.

Change-Id: I93d46e11a4245288f0e87c87a2e4bf45ac5aff69

build: Add changes to release tools and mkbootimg

Change the prototype of LoadRecoveryFstab to take in the device type
as a argument. Fix the case where mkbootimg was being passed a incorrect
argument.

Change-Id: Ic6ac596d8d96d3a5effbdf513763ec1cb92b1a030