build: add kernel header dependency if module uses kernel headers
Many of the QCOM components use kernel headers, but don't declare
the dependency on them.  This is fine in CAF because of the way they
build the boot.img before anything else.  In CM, we don't build the
boot.img the same, so we run into a race between the kernel build &
these modules... and the modules lose.

Warn about modules that have this missing dependency, and add it for
them so we don't have to modify each Android.mk.

Change-Id: I95f1e47b5ef440f6f5d8f64a0c3f38d9572e839e
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>

build: Warn on deprecated kernel header path

The build is switching to the target INSTALLED_KERNEL_HEADERS to
declare dependencies on kernel headers.  Warn when the build
encounters the old style dependency $(KERNEL_OUT)/usr.

Change-Id: I913e74681b02dfcf1eaed3d1e47ff4ab2300b12d
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>

Remove deprecated kernel header dependency path warning

Change-Id: I68b44fd1e868a3326a6ab1a7e5fe42b8eeecd1c1

Don't error on kernel headers in LOCAL_ADDITIONAL_DEPENDENCIES

The new kernel header dependency path is wrongly picked up by this
logic. Check if KERNEL_OBJ is contained in the string and only error
if it isn't.

Change-Id: I5425f11eb8b650a80ce1dda303514fdc64888af9

build: Automatically replace old-style kernel header includes with new header lib

Since we do this via soong and the result is a header library, just replace all
legacy include paths that point to the old header location with calls to the new
header library. Since we no longer have the legacy include, we can also remove
the additional dependency.

This reverts commit fa798218e5b7f96e12ab3acc9d47a3f26a140777.

Change-Id: I716955534e50831e6568ca01e480aa8b90075d92

build: Also replace techpack/audio old-style kernel header

 * We are also going to replace them by the new header lib

Change-Id: I0562d8f0cfe3186af50e00eaab858f2836cbc9f0